vp_abstractor/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vp-abstractor"
version = "0.11.0"
description = "."
requires-python = ">=3.8"
dependencies = [
    "kfp>=2.13.0",
    "google-cloud-aiplatform>=1.86.0",
    "google-cloud-storage>=2.19.0",
    "google-cloud-pipeline-components>=2.20.0",
    "google-auth>=2.38.0",
    "docker>=7.1.0",
    "google-cloud-monitoring>=2.20.0"
]